Adds a button with a timer
Creates a button with an on click event that increases the width of the button. A timer based on the width of the button will decrease the width. The number on the left of the button shows the current width while the number on the right shows the highest the width has been. My personal best without cheating is 845. Good luck!